SERVER-5253 Fork mongodb files that comprise the C++ client driver.
This is a subset of the files found in the mongodb server git repository
found here: https://github.com/mongodb/mongo

These files were forked from the 2.6.0-rc0 tag:
https://github.com/mongodb/mongo/tree/r2.6.0-rc0

The forked files will be used to provide a C++ client driver that
can be maintained, versioned, and released independently of the server. The
files have been imported on a 'legacy' branch: our long term plan is to
provide a new C++ driver, and the 'master' branch is saved for that work.