CXX-151 As part of RC3 backport, drop signal_handlers.
These files changed significantly between rc2 and rc3, including
pulling in a new dependency on db/client.h, which we had not extracted.

We def. don't want db/client.h and its transitive dependencies, but
as a nice consequence of other edits in rc2..rc3, no files that we
own now depend on signal_handlers.h, so we can drop the dependency,
rather than needing to pull in db/client.h

Note that we have already made this change on the legacy branch. The
driver library shouldn't be in the business of signal handling.