CXX-717: implement C++ version verification
mongo-cxx-driver crashes if used in an application built with a
different C++ version than the library itself. Add a little compile time
generator that adds the library C++ version and a snippet in config.h
that throws an error if compiling a file against mongo-cxx-driver
headers with a different C++ version.