CXX-1000 Migrate documentation to GitHub Pages
This commit creates a docs/ directory containing a general set
of documentation files for use with the hugo static site generator.
The site generated from these files will be deployed to GitHub Pages
and supersede the current wiki.

It also provides CMake targets to generate and deploy both this static
manual and the doxygen generated API documentation sites for all legacy and
mongocxx releases.

The doxygen configurations are generated dynamically to ensure that
documentation from old releases uses the same formatting and "front page"
content as current releases.

The site deployment targets do not create symlinks, which are necessary for
proper "permalinks" to the latest documentation.  These must be added
manually:

* /api/mongocxx-v3 -> /api/mongocxx-3.0.x (for latest x in v3 series)
* /api/legacy-v1   -> /api/legacy-1.1.x (for latest x in v1 series)
* /api/current     -> /api/mongocxx-v3

For symlinks to work, there must be a `.nojekyll` file at the top of
the GitHub Pages site.