CXX-1045 Add hint::to_value(), deprecate hint::to_document()
The design of hint::to_document() is flawed in that it returns a
document with a fixed key name of "$hint", which doesn't always
represent the hint option correctly for database operations (commands
take this option with a key name of "hint", for example).

To address this issue, this commit removes knowledge of the key name
from the hint class, by adding hint::to_value() and deprecating
hint::to_document().