CXX-999 Add pipeline::view_array(), deprecate pipeline::view()
This commit provides a mechanism to get the underlying array out of a
pipeline object, which logically wraps an array builder.  This is
needed for database commands which take a pipeline option as a BSON
array type.

The pipeline::view() method exposes a document wrapper around this
array, which is strictly less useful.