CXX-1001 CXX-1100 Fix mongocxx cursor iteration
CXX-1001 reported the repeated calls to mongocxx::cursor::begin() would
increment the cursor and lose documents.  This is fixed by only
incrementing the cursor during construction one time (required to send the
query/command to the server).

CXX-1100 reported that incrementing an iterator would leave all other
iterators on that cursor with dangling pointers.  While this is
technically legal, if undefined behavior, for input iterators, the team
decided to have all iterators from the same cursor move in lockstep;
this is also legal and prevents undefined behaviors, which is more
user-friendly.  This is achieved by moving the "current document" view
into the private cursor implementation rather than keeping copies in
individual iterators.

The changes above required some additional coordination to track cursors
sent and dead/finished.  Strangely, we couldn't depend on libmongoc
functions for this due to what appear to be some inconsistencies in
libmongoc's handling of "traditional" cursors and "command cursors".