CXX-1167 Fix broken SSL testing
The unit tests to confirm that a pool could be configured with SSL
options never ran because the tests were guarded by a private definition
that was never included. This also concealed a bug where mocking failed
because the necessary private header wasn't included in a test helper header.

This commit adds the private header to the test helper header to fix
mocking and runs the SSL pool configuration test unconditionally if
MONGOC_ENABLE_SSL is configured.  I manually verified that this commit
would have caught the original CXX-1163 bug.

Backported from CXX-1163.