When patching include paths to use 'mnmlstc', tell perl to avoid filename collisions.
By default, perl's "-i" option tells it to:
* Open the specified file.
* Unlink the file.
* Create a new file with the same name.
* Copy the old file to the new file, transforming it per the regex.

This is not valid on Windows because you can't have two files with the
same name, even if one of them has been unlinked.

The quick fix is to tell Perl to instead create a backup file, so it
moves the original file rather than unlinking it.

(Amended by xdg to clean up .bak files)