CXX-1257 Add export macros to some private syms when building test lib
This work adds new macros BSONCXX_TEST_API and MONGOCXX_TEST_API,
which are intended to be used to tag private functionality which needs
to be referenced by the driver tests.

This is groundwork for allowing the tests to link against the shared
library.  Currently, the tests unconditionally link to the static
library.