cmake: honor LIB_SUFFIX
On some systems, e.g., Fedora, libraries are installed to /usr/lib64.
A typical pattern is to honor a variable named LIB_SUFFIX. It is common
and specified by default in the Fedora build system if necessary.

This patch extends the installation paths from "lib" to
"lib${LIB_SUFFIX" in order to allow for proper packaging and
installation.

The suffix can be set as "-DLIB_SUFFIX=64" and is ignored if not set.