Fix clang lambda capture error
chunk_size is a constexpr so it need not be in the capture list. At the
same time, if chunk_size is not in the capture list then MSVC complains
'chunk_size' cannot be implicitly captured because no default capture
mode has been specified. This capture spec seems to make both happy.