CXX-1901 check for old libbson cmake pkg name
CDRIVER-3047 added new cmake package names for the C driver
(bson-1.0/mongoc-1.0) and deprecated the old ones (libbson-1.0/
libmongoc-1.0 and libbson-static-1.0/libmongoc-static-1.0).

CXX-1707 changed the C++ driver to use the new names of the
C driver. That implicitly requires the latest C driver to build
against. So, this change fall backs to the old package names
if the new ones are not found.