Create docker images for C++ driver
This creates a CI build and test for docker images based on the
following distros:
- Alpine
- Debian
- RHEL

Each distro version is in its own directory. This directory contains a
Makefile, a test C++ program to ensure that linking works properly, and
a TestDockerfile to ensure that the image was built properly.

To build an image locally, navigate to the directory of the
distro/version that you wish to build, then type:
```
make
```