Bump required C driver to 1.25.0 (#1056)
* bump required C driver to 1.25.0

* remove reference to removed EVG variant

"Ubuntu 18.04 with minimum libmongoc" was removed in PR #1049. Testing with minimum libmongoc is now the default.

* remove step for updating docker versions

This step is listed in later in `Docker Image Build and Publish`.

Building the Docker images requires downloading the C++ release tarball. Bump after tagging.

* remove `check_docker_file_versions`

The Docker versions are bumped after tagging the release. This check may not be expected to succeed during a release.

* fix documented introduction of the auto C driver download

It is introduced in 3.9.0 as part of CXX-2695.

* add `-pthread` to example project build scripts

This is a workaround until CDRIVER-4776 is resolved.