Update release instructions (#1424)
* fix `sed` command on macOS

* remove unnecessary heading

The `pre-release-changes` branch is only created for the non-patch release steps. Combine the step within non-patch release steps.

* change "will" to "may"

A new branch is only created for a non-patch release.

* use X.Y

For consistency with other version placeholders

* fix link in documentation changes needed

Previous link is not published for newer releases after (latest is 3.10.0).

* do not include next patch release on master branch

I do not expect upcoming patch release notes will be added on the master branch. Patch releases are tracked on release branches.

---------

Co-authored-by: Ezra Chung <88335979+eramongodb@users.noreply.github.com>